OB-2616: Initial capture and playback implementation for signals
This change adds a wrapper for recording signals in flight to disk so
that they can be later replayed. It also contains a playback
implementation.

I have tested this and it seems to work, but playback does not work as
expected. I believe it makes sense for us to pair at this point because
I think I'm running into an understanding issue.

Future work still needed:
* Potentially the playback should be keyed on both signal type and
signal_id.
* Be able to handle the same signal being used multiple times
* Fix the playback because it doesn't seem to work as I expect when
testing against onebook