RFLX-3: Add a metric for error/pending/successful GraphQl queries
The current metrics require some metric manipulation to get an estimate
of the state of the world with respect to errors. This change introduces
specific metrics for the current status.

Caveat - this is an unoptimised and potentially slow implementation. We
have options for making this smarter and lighter weight if there's need.